+++ Bugsnag/BugsnagSystemState.m
@@ -97,6 +97,7 @@ static NSMutableDictionary* initCurrentState(BugsnagKVStore *kvstore, BugsnagCon
     app[BSGKeyReleaseStage] = config.releaseStage;
     app[BSGKeyVersion] = blankIfNil(systemInfo[@BSG_KSSystemField_BundleShortVersion]);
     app[BSGKeyBundleVersion] = blankIfNil(systemInfo[@BSG_KSSystemField_BundleVersion]);
+    app[BSGKeyMachoUUID] = [BSG_KSSystemInfo appUUID];
     app[@"inForeground"] = @(isInForeground);
     app[@"isActive"] = @(isActive);
 #if BSG_PLATFORM_TVOS
@@ -159,7 +160,7 @@ static NSDictionary *copyDictionary(NSDictionary *launchState) {
 @implementation BugsnagSystemState
 
 - (instancetype)initWithConfiguration:(BugsnagConfiguration *)config {
-    if (self = [super init]) {
+    if ((self = [super init])) {
         _kvStore = [BugsnagKVStore new];
         _persistenceFilePath = [BSGFileLocations current].systemState;
         _lastLaunchState = loadPreviousState(_kvStore, _persistenceFilePath);
@@ -235,11 +236,6 @@ static NSDictionary *copyDictionary(NSDictionary *launchState) {
     [self mutateLaunchState:^(NSMutableDictionary *state) {
         state[BSGKeySession] = notification.object;
     }];
-}
-
-- (void)recordAppUUID {
-    // [BSG_KSSystemInfo appUUID] returns nil until we have called _dyld_register_func_for_add_image()
-    [self setValue:[BSG_KSSystemInfo appUUID] forAppKey:BSGKeyMachoUUID];
 }
 
 - (void)setCodeBundleID:(NSString*)codeBundleID {