+++ Bugsnag/KSCrash/Source/KSCrash/Recording/BSG_KSCrashIdentifier.h
@@ -6,19 +6,18 @@
 extern "C" {
 #endif
 
-void bsg_kscrash_generate_report_initialize(const char *directory, const char *bundleName);
+void bsg_kscrash_generate_report_initialize(const char *directory);
 
 /**
  * Generates a new UUID. Not async signal safe. Caller responsible for
  * freeing allocated string.
  */
-const char *bsg_kscrash_generate_report_identifier(void);
+char *bsg_kscrash_generate_report_identifier(void);
 /**
  * Generates a new path string. Not async signal safe. Caller responsible
  * for freeing allocated string.
  */
-const char *bsg_kscrash_generate_report_path(const char *identifier,
-                                             bool is_recrash_report);
+char *bsg_kscrash_generate_report_path(const char *identifier, bool is_recrash_report);
 
 #ifdef __cplusplus
 }