+++ Bugsnag/Payload/BugsnagApp.m
@@ -37,7 +37,7 @@ NSDictionary *BSGParseAppMetadata(NSDictionary *event) {
         app.releaseStage = json[@"releaseStage"];
         app.type = json[@"type"];
         app.version = json[@"version"];
-        app.dsymUuid = json[@"dsymUUIDs"][0];
+        app.dsymUuid = [json[@"dsymUUIDs"] firstObject];
     }
     return app;
 }
@@ -60,13 +60,28 @@ NSDictionary *BSGParseAppMetadata(NSDictionary *event) {
           codeBundleId:(NSString *)codeBundleId
 {
     NSDictionary *system = event[BSGKeySystem];
-    app.id = system[@"CFBundleIdentifier"];
-    app.bundleVersion = config.bundleVersion ?: system[@"CFBundleVersion"];
-    app.dsymUuid = system[@"app_uuid"];
-    app.version = config.appVersion ?: system[@"CFBundleShortVersionString"];
-    app.releaseStage = config.releaseStage;
+    app.id = system[@BSG_KSSystemField_BundleID];
+    app.bundleVersion = system[@BSG_KSSystemField_BundleVersion];
+    app.dsymUuid = system[@BSG_KSSystemField_AppUUID];
+    app.version = system[@BSG_KSSystemField_BundleShortVersion];
     app.codeBundleId = [event valueForKeyPath:@"user.state.app.codeBundleId"] ?: codeBundleId;
-    app.type = config.appType;
+    [app setValuesFromConfiguration:config];
+}
+
+- (void)setValuesFromConfiguration:(BugsnagConfiguration *)configuration
+{
+    if (configuration.appType) {
+        self.type = configuration.appType;
+    }
+    if (configuration.appVersion) {
+        self.version = configuration.appVersion;
+    }
+    if (configuration.bundleVersion) {
+        self.bundleVersion = configuration.bundleVersion;
+    }
+    if (configuration.releaseStage) {
+        self.releaseStage = configuration.releaseStage;
+    }
 }
 
 - (NSDictionary *)toDict